增加注释

+++ WindowManager/CustomGraphics/src/CustomScene.cpp
@@ -123,7 +123,9 @@ const QRectF CustomScene::getSelectRect()
         qDebug() << "SelectRect or ImageItem not found in the scene.";
         return QRectF();
     }
-    float inverseScale = 1 / pImgItem->scale();
+    
+    // 由于是在缩放后的图像上进行选择，所以需要将在缩放后的选择矩形转换回原图像的坐标系
+    qreal inverseScale = 1 / pImgItem->scale();
 
     return QRectF(pImgItem->mapFromItem(pSelectItem, pSelectItem->boundingRect().topLeft()),
                   pSelectItem->boundingRect().size() * inverseScale);
